backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.17.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==1.3.0
pytest-cov==6.0.0
ruff==0.8.5
black==25.11.0
slowapi==0.1.9
anthropic==0.39.0
openai==1.59.5
